Add (Hidden)Forfeit Button. Extract Alert component
