+++ client/src/components/index.js
@@ -1,3 +1,4 @@
+export { default as Alert } from './Alert/Alert'
 export { default as Avatar } from './Avatar/Avatar'
 export { default as Button } from './Button/Button'
 export { default as Card } from './Card/Card'
