+++ client/src/pages/OpponentSelect/components/index.js
@@ -1,3 +1,2 @@
-export { default as ActiveBattleAlert } from './ActiveBattleAlert/ActiveBattleAlert'
 export { default as BattlePreviewModal } from './BattlePreviewModal/BattlePreviewModal'
 export { default as OpponentCard } from './OpponentCard/OpponentCard'
